12 point star coil template
A little idea I was experimenting with, i mess with electronics a lot, and before programming I was fascinated with electromagnetism. I found that a typical coil (cylindrical) will generate a uniform field when current is passed through it - providing the coil is symmetrically wound, but by changing the shape of the coil you change the shape of the EM field. This star coil is a variation of Rodin's coil designs (a toroid or donut shaped coil).

I found these coils to have some very peculiar properties. When a VERY small current is fed via its windings, a VERY strong EM field is generated at the center cavity. I can only describe it as taking 2 super strong magnets, and feeling the push effect as the same poles repel eachother. This coil design replicates that, but in a uniform circle creating a heavy force at the center. The result? If you put a metal ball-bearing in the center and supply just 1.5 Volts DC (a single AA battery) the bearing is propelled like a projectile at great